break up useWallets
