+++ src/hooks/useWallets.ts
@@ -1,12 +1,9 @@
 import { useMemo } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { createSelector } from 'reselect';
 import { findLatestBackUp } from '../model/backup';
-import { addressSetSelected, walletsSetSelected } from '../redux/wallets';
-import useInitializeWallet from './useInitializeWallet';
-import { toChecksumAddress } from '@/handlers/web3';
 import WalletTypes from '@/helpers/walletTypes';
-import { RainbowAccount, RainbowWallet } from '@/model/wallet';
+import { RainbowWallet } from '@/model/wallet';
 import { AppState } from '@/redux/store';
 import logger from '@/utils/logger';
 
@@ -27,8 +24,6 @@ const walletSelector = createSelector(
 );
 
 export default function useWallets() {
-  const initializeWallet = useInitializeWallet();
-  const dispatch = useDispatch();
   const { isWalletLoading, latestBackup, selectedWallet, walletNames, wallets } = useSelector(walletSelector);
 
   const isDamaged = useMemo(() => {
@@ -41,20 +36,6 @@ export default function useWallets() {
     return bool;
   }, [selectedWallet, wallets]);
 
-  const switchToWalletWithAddress = async (address: string): Promise<string | null> => {
-    const walletKey = Object.keys(wallets!).find(key => {
-      // Addresses
-      return wallets![key].addresses.find((account: RainbowAccount) => account.address.toLowerCase() === address.toLowerCase());
-    });
-
-    if (!walletKey) return null;
-    const p1 = dispatch(walletsSetSelected(wallets![walletKey]));
-    const p2 = dispatch(addressSetSelected(toChecksumAddress(address)!));
-    await Promise.all([p1, p2]);
-    // @ts-expect-error ts-migrate(2554) FIXME: Expected 8-9 arguments, but got 7.
-    return initializeWallet(null, null, null, false, false, null, true);
-  };
-
   return {
     isDamaged,
     isReadOnlyWallet: selectedWallet.type === WalletTypes.readOnly,
@@ -62,7 +43,6 @@ export default function useWallets() {
     isWalletLoading,
     latestBackup,
     selectedWallet,
-    switchToWalletWithAddress,
     walletNames,
     wallets,
   };
